Additional updates to the go test code
The key change is the introduction of a retry loop when attempting to
call LinkSubscribeAt on a new network namespace.  It seems that it takes
a little while before a new network namespace is ready for subscribing.
On my system, it take about 50 usecs (give or take a few usecs).

Signed-off-by: Andre Fredette <afredette@redhat.com>